Pass II original functionality complete

+++ Assembler-VC407-Emulator/Emulator.h
@@ -14,7 +14,7 @@ public:
          m_memory.resize(MEMSZ, 0);
     }
     // Records instructions and data into simulated memory.
-    bool insertMemory( int a_location, long long a_contents );
+    bool insertMemory( int a_location, int a_contents );
     
     // Runs the program recorded in memory.
     bool runProgram( );
